requirements.txt
asgiref
crispy-bootstrap5
dj-database-url  
Django
django-crispy-forms       
django-environ
django-filter
djangorestframework       
Markdown
MouseInfo
pillow
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
sqlparse
typing_extensions
tzdata
